Error handling refactored in the api routes "/[issueId]", "/issues" and "./statusChanges".

+++ pages/api/issues/[issueId].js
@@ -1,24 +1,42 @@
 import prisma from "@/prisma/client";
-import { NextApiHandler, NextApiResponse } from "next";
+import { getServerSession } from "next-auth/next";
+import { authOptions } from "../auth/[...nextauth]";
 
 export default async function handler(req, res) {
-  console.log("REQ BODY->", req.body);
-
-  // const issueId = req.query.id;
   const issueId = req.query.issueId;
-  console.log("from handler", issueId);
+
+  const session = await getServerSession(req, res, authOptions);
 
   if (req.method === "DELETE") {
-    await prisma.statusChange.deleteMany({ where: { issueId: issueId } });
+    if (!session) {
+      return res.status(401).end();
+    }
+
+    const issue = await prisma.issue.findUnique({
+      where: {
+        id: issueId,
+      },
+    });
+
+    if (!issue) {
+      return res.status(404).end();
+    }
+
+    const deletedChanges = await prisma.statusChange.deleteMany({
+      where: { issueId: issueId },
+    });
+
     const issueDelete = await prisma.issue.delete({
       where: {
         id: issueId,
       },
     });
+
     return res.status(200).json({ issueDelete });
   } else {
-    return res
-      .status(405)
-      .json({ message: `Method ${req.method} not supported` });
+    res.status(405).end();
   }
 }
+
+//Maybe used as reference?
+//  //console.log("Session", JSON.stringify(session, null, 2));
